# Reduce some allocations in Razor