--- a/src/Razor/Microsoft.AspNetCore.Razor.Language/src/Legacy/LegacySyntaxNodeExtensions.cs
+++ b/src/Razor/Microsoft.AspNetCore.Razor.Language/src/Legacy/LegacySyntaxNodeExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -48,11 +48,6 @@ namespace Microsoft.AspNetCore.Razor.Language.Legacy
 
         public static SpanContext GetSpanContext(this SyntaxNode node)
         {
-            if (node == null)
-            {
-                throw new ArgumentNullException(nameof(node));
-            }
-
             var context = node.GetAnnotationValue(SyntaxConstants.SpanContextKind);
 
             return context is SpanContext ? (SpanContext)context : null;