… src/Razor/Microsoft.AspNetCore.Razor.Language/src/DefaultRazorTagHelperBinderPhase.cs
… src/Razor/Microsoft.AspNetCore.Razor.Language/src/StringSegment.cs
… src/Razor/Microsoft.AspNetCore.Razor.Language/src/TagHelperDocumentContext.cs
… src/Razor/Microsoft.AspNetCore.Razor.Language/src/TagHelperMatchingConventions.cs
… src/Razor/Microsoft.AspNetCore.Razor.Language/test/DefaultRazorTagHelperBinderPhaseTest.cs
… src/Razor/Microsoft.AspNetCore.Razor.Language/test/StringSegmentTest.cs