Nothing particularly dramatic, just some low hanging fruits that remove string allocations
in TagHelperMatchingConvention. VS ships with a copy of System.Memory, so we're not the first to
introduce this dependency over there.

Some before and after benchmarks because there was a convenient scenario set up for this:

// * Summary *

BenchmarkDotNet=v0.12.1.1466-nightly, OS=Windows 10.0.19043
Intel Core i7-8700 CPU 3.20GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK=6.0.100-preview.5.21230.2
  [Host]     : .NET Core 2.1.27 (CoreCLR 4.6.29916.01, CoreFX 4.6.29916.03), X64 RyuJIT
  Job-GISAGT : .NET Core 2.1.27 (CoreCLR 4.6.29916.01, CoreFX 4.6.29916.03), X64 RyuJIT

Server=True  Toolchain=.NET Core 2.1  RunStrategy=Throughput

Before:
|                                  Method |        Mean |     Error |    StdDev |     Op/s |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|---------------------------------------- |------------:|----------:|----------:|---------:|-------:|-------:|-------:|----------:|
|      'TagHelper Design Time Processing' | 2,313.36 us | 25.477 us | 21.275 us |    432.3 | 7.8125 | 3.9063 | 3.9063 |    986 KB |
| 'TagHelper Component Directive Parsing' |    93.52 us |  1.408 us |  1.100 us | 10,693.5 |      - |      - |      - |      3 KB |

Ater:
|                                  Method |        Mean |     Error |    StdDev |     Op/s |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|---------------------------------------- |------------:|----------:|----------:|---------:|-------:|-------:|-------:|----------:|
|      'TagHelper Design Time Processing' | 2,205.42 us | 28.353 us | 26.522 us |    453.4 | 7.8125 | 3.9063 | 3.9063 | 963 KB |
| 'TagHelper Component Directive Parsing' |    85.89 us |  0.724 us |  0.565 us | 11,643.1 |      - |      - |      - | 288 B |

Also some minor cleanup of Linq operations and null-ref checks in internal code


<!-- Thank you for submitting a pull request to our repo. -->

<!-- If this is your first PR in the ASP.NET Core repo, please run through the checklist
below to ensure a smooth review and merge process for your PR. -->

- [ ] You've read the [Contributor Guide](https://github.com/dotnet/aspnetcore/blob/main/CONTRIBUTING.md) and [Code of Conduct](https://github.com/dotnet/aspnetcore/blob/main/CODE-OF-CONDUCT.md).
- [ ] You've included unit or integration tests for your change, where applicable.
- [ ] You've included inline docs for your change, where applicable.
- [ ] There's an open issue for the PR that you are making. If you'd like to propose a new feature or change, please open an issue to discuss the change or find an existing issue.

<!-- Once all that is done, you're ready to go. Open the PR with the content below. -->

**PR Title**
Summary of the changes (Less than 80 chars)

**PR Description**
Detail 1
Detail 2

Addresses #bugnumber (in this specific format)